componentize comments
